Move vision subsystem to robotcontainer Fixes #5
